pyproject.toml
[project]
name = "kayan-constructions"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohamed Ismail",email = "mohamed.g.ismail@orange.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-imagekit (>=5.0.0,<6.0.0)",
    "django-colorfield (>=0.14.0,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
